detect/http-client-body: move tests into tests/
