+++ src/detect.c
@@ -54,7 +54,6 @@
 
 #include "detect-engine-payload.h"
 #include "detect-engine-event.h"
-#include "detect-engine-hcbd.h"
 #include "detect-engine-hsbd.h"
 
 #include "detect-filestore.h"
